feat(#3) :: 사이드바 완료 & 투표#1
