+++ apps/client/src/app/page.tsx
@@ -1,7 +1,7 @@
 export default function Home() {
   return (
-    <div className="vote_main">
-      <h3>안녕</h3>
+    <div >
+      <h3>투표로 만들어가는 가이드, UP!K입니다</h3>
     </div>
   );
 }
